+++ src/date-picker/examples/basic/basic.component.ts
@@ -12,9 +12,16 @@ export class ThyDatePickerBasicExampleComponent implements OnInit {
     } as DateEntry;
 
     date = { date: new Date(), with_time: 0 };
+
     dateTime = 1234567890;
 
+    week = { date: new Date(), with_time: 0 };
+
     dateRange = { begin: new Date('2021-10'), end: new Date('2021-12') };
+
+    weekRange = { begin: new Date('2021-10-03'), end: new Date('2021-12-12') };
+
+    yearRange = { begin: new Date('2021'), end: new Date('2028') };
 
     flexibleDateRange: ThyDateRangeEntry;