+++ src/watermark/examples/module.ts
@@ -5,15 +5,6 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ThyInputModule } from 'ngx-tethys/input';
 
-import { ThyWatermarkBasicExampleComponent } from './basic/basic.component';
-import { ThyWatermarkCanvasCustomExampleComponent } from './custom/custom.component';
-
-const COMPONENTS = [ThyWatermarkBasicExampleComponent, ThyWatermarkCanvasCustomExampleComponent];
-
-@NgModule({
-    declarations: COMPONENTS,
-    imports: [CommonModule, FormsModule, ThyWatermarkModule, ThyInputModule],
-    exports: COMPONENTS,
-    providers: []
-})
-export class ThyWatermarkExamplesModule {}
+export default {
+    imports: [CommonModule, FormsModule, ThyWatermarkModule, ThyInputModule]
+};